exception handlers added for findbyID and deleteByID for Console and TShirts

+++ src/main/java/com/company/Summative1CristieJBNicholas/controller/ConsoleController.java
@@ -10,6 +10,10 @@ import org.springframework.web.bind.annotation.*;
 import java.util.List;
 import java.util.Optional;
 
+
+// using JSR 303 annotations to validate inputs
+
+// create a global error ahndler for all exceptions thrown by controller methods
 @RestController
 public class ConsoleController {
     @Autowired
@@ -30,9 +34,13 @@ public class ConsoleController {
         return repo.findAll();
     }
 
-    // find Console by iD
+    // find Console by iD; throws 422 error if invalid ID is selected
     @GetMapping("/Console/consoleId/{consoleId}")
     public Console getConsoleById(@PathVariable Integer consoleId) {
+    // from echo-range-service class work
+        if (consoleId < 1) {
+            throw new IllegalArgumentException("Console ID must be at least 1");
+        }
         Optional<Console> returnVal = repo.findById(consoleId);
         return returnVal.get();
     }
@@ -40,11 +48,16 @@ public class ConsoleController {
     // find Console by Manufacturer
     @GetMapping("/Console/manufacturer/{manufacturer}")
     public List<Console> getConsolesByManufacturer(@PathVariable String manufacturer) {
+
+        // need to find better logic to deal w/ string (don't use  == null)
+        if ((manufacturer == null)) {
+            throw new IllegalArgumentException("Must choose a manufacturer");
+        }
         return repo.findAllConsolesByManufacturer(manufacturer);
     }
 
-   // update an existing Console record
-    @PutMapping(value="/Console")
+    // update an existing Console record
+    @PutMapping(value = "/Console")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void updateConsole(@RequestBody Console console) {
         repo.save(console);
@@ -55,6 +68,9 @@ public class ConsoleController {
     @DeleteMapping("/Console/{consoleId}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void deleteConsole(@PathVariable Integer consoleId) {
+        if (consoleId < 1) {
+            throw new IllegalArgumentException("Cannot delete a Console unless Console ID is at least 1");
+        }
         repo.deleteById(consoleId);
     }
 }
